Initial commit of the ultimate fall back data structure.  The basic stuff all works (via ppss|btshell then "+ 1\n+ 2\np\n", say).  Much work remains to be done to round out the whole impl.  Pertinent discussions surrounding this are at:     https://github.com/nim-lang/Nim/pull/13890     https://github.com/nim-lang/RFCs/issues/207     https://github.com/nim-lang/fusion/commit/f900468fd3ee568e5cfa000da9508b3d94b53ceb
Future downstream applications of this are a fallback from either abused
or attacked hash tables or `seq`s (many in-the-middle/deep inserts).
Both such fallbacks are in contexts where callers (implicitly) already
accept "lumpy" O(N) performance in the middle of the offending operation
(either a hash resize or the shifting in the middle for seq).